Change PlanPerHour/Version to decimal, improve product UI
- Changed PlanPerHour and Version fields from int? to decimal? across backend, API, and frontend to support fractional values.
- Added EF Core migration to update PRODUCTS table columns to numeric.
- ProductStatus enum now mapped between domain, API, and client; added Russian-friendly status display and selection.
- SKU is now generated server-side on product creation.
- Refactored RemoveProductMaterial API to use ProductId and MaterialId.
- Major UI/UX improvements for product details and edit pages, including better field display, status localization, and error messages.
- Enhanced BOM editing logic and duplicate checks in ProductEditPageViewModel.
- Products list now shows SKU and status columns.
- Cleaned up code, improved error handling, and updated API examples.

+++ src/MyFactory.Application/DTOs/Products/ProductDetailsDto.cs
@@ -5,9 +5,9 @@ public sealed record ProductDetailsDto
     public Guid Id { get; init; }
     public string Sku { get; init; } = default!;
     public string Name { get; init; } = default!;
-    public int? PlanPerHour { get; init; }
+    public decimal? PlanPerHour { get; init; }
     public string? Description { get; init; }
-    public int? Version { get; init; }
+    public decimal? Version { get; init; }
     public Domain.Entities.Products.ProductStatus Status { get; init; }
 
     public decimal MaterialsCost { get; init; }
