+++ src/MyFactory.WebApi/Contracts/Products/UpdateProductRequest.cs
@@ -4,7 +4,7 @@ namespace MyFactory.WebApi.Contracts.Products;
 
 public record UpdateProductRequest(
     string Name,
-    int? PlanPerHour,
+    decimal? PlanPerHour,
     ProductStatus Status,
     string? Description,
-    int? Version);
+    decimal? Version);
